fix: refactoring - implemented environment variables on .env file and application.yml
