+++ .env
@@ -1,17 +1,17 @@
-db_name=ms-ml-service
-db_timezone=America/Sao_Paulo
-db_version=0.0.4
-db_driver=mongodb+srv://
-db_cluster=cluster0.otugy2g.mongodb.net/
-db_user=herdedbclusteriros_aurora
-db_pass=herdeiros_aurora_001
-db_name1=ml_service_non_handler_db
-db_name2=dev_ml_service_db
-db_uri1=http://201.95.98.141
-db_port1=8081
-db_uri2=http://201.95.98.141
-db_port2=8082
-db_url=https://valorant-api.com/v1
+MS_NAME=ms-ml-service
+TZ=America/Sao_Paulo
+MS_VERSION=0.0.5
+MONGO_DRIVER=mongodb+srv://
+MONGO_CLUSTER=cluster0.otugy2g.mongodb.net/
+MONGO_USER=herdedbclusteriros_aurora
+MONGO_PASS=herdeiros_aurora_001
+MONGO_DB_NAME_NON=ml_service_non_handler_db
+MONGO_DB_NAME=dev_ml_service_db
+MS_GATEWAY_URI=http://201.95.98.141
+MS_GATEWAY_PORT=8081
+MS_BASE_SERVICE_URI=http://201.95.98.141
+MS_BASE_SERVICE_PORT=8082
+VALORANT_URI=https://valorant-api.com/v1
