+++ batches/valorant/main.py
@@ -1,20 +1,17 @@
 # -*- coding: utf-8 -*-
-
+import json
 import os
 import sys
-import urllib as ur
-import json
-import yaml
+import urllib.request as ur
 
-# set path sys default root 
 sys.path.append(os.path.abspath(os.getcwd()))
 
-# import dependencies local
 from services.pymongo_service import PymongoService
 from domain.valorant.model.agents_model import AgentsModel
-from services.log_service import LogService
+from services.log.log_service import LogService
+from services.env.build_env_service import BuildEnvService
 
-    # storage path root and some constants variables
+# storage path root and some constants variables
 AGENTS = 'agents'
 VALORANT = 'valorant'
 DOMAIN = 'domain'
@@ -28,6 +25,7 @@ VALORANT_PATH = f'{STORAGE_PATH}/{VALORANT}'
 VALORANT_FILE_PATH = f'{VALORANT_PATH}/files'
 VALORANT_IMAGE_PATH = f'{VALORANT_PATH}/images'
 
+
 def insert_data(db, collection_name, data) -> None:
     """
     description:
@@ -38,9 +36,10 @@ def insert_data(db, collection_name, data) -> None:
     see: 
         https://www.mongodb.com/pt-br/docs/manual/reference/method/db.collection.insert/
     """
-    
+
     LogService.log().info(f'[batch valorant] (insert_data): insert into database')
     db[collection_name].insert_one(vars(data))
+
 
 def build_model_valorant(data) -> AgentsModel:
     """
@@ -56,7 +55,7 @@ def build_model_valorant(data) -> AgentsModel:
     return AgentsModel(json.dumps(json.loads(data)))
 
 
-def getData(url: str) -> str:
+def get_data(url: str) -> str:
     """
     description:
         this method is responsible to make http call on api valorant and get some data
@@ -66,14 +65,14 @@ def getData(url: str) -> str:
         http response from valorant and convert to string
     """
     LogService.log().info(f'[batch valorant] (getData): request open url')
-    response = ur.request.urlopen(url)
-    
+    response = ur.urlopen(url)
+
     LogService.log().info(f'[batch valorant] (getData): read reponse http')
     byte_data = response.read()
-    
+
     LogService.log().info(f'[batch valorant] (getData): convert chartset to utf-8')
     encoding = response.info().get_content_charset('utf8')
-    
+
     LogService.log().info(f'[batch valorant] (getData): return data string')
     return byte_data.decode(encoding)
 
@@ -84,45 +83,31 @@ def create_dir():
         this method is responsible to create some dirs if not extis
     """
     LogService.log().info(f'[batch valorant] (create_dir): create some dirs if not exists')
-    for dir in [VALORANT_PATH, VALORANT_FILE_PATH, VALORANT_IMAGE_PATH]:
-        if not os.path.exists(dir):
-            os.mkdir(dir)
-
-
-def config():
-    """
-    description:
-        this method is responsible to load some variables from yaml file located in root path
-    """
-    LogService.log().info(f'[batch valorant] (config): read variables from yaml file')
-    with open(f'{ROOT_PATH}/application.yml') as file:
-        root = yaml.safe_load(file)
-        file.close()
-
-    assert root
-    return root
+    for directory in [VALORANT_PATH, VALORANT_FILE_PATH, VALORANT_IMAGE_PATH]:
+        if not os.path.exists(directory):
+            os.makedirs(directory, exist_ok=True)
 
 
 def initialize():
     LogService.log().info(f'[batch valorant] (initialize): check connection database')
     assert PymongoService().get_mongo_client().admin.command('ping')
-    
+
     LogService.log().info(f'[batch valorant] (initialize): load some variables config')
-    root = config()
-    
-    valorant = root.get(DOMAIN).get(VALORANT)
-    mongo = root.get(INFRA).get(MONGO)
+    env = BuildEnvService().load_environment(f'{ROOT_PATH}/application.yml')
+
+    valorant = env.get(DOMAIN).get(VALORANT)
+    mongo = env.get(INFRA).get(MONGO)
 
     LogService.log().info(f'[batch valorant] (initialize): create some dirs if not exists')
     create_dir()
-    
+
     LogService.log().info(f'[batch valorant] (initialize): loading instance mongodb')
     mongo_instance = PymongoService.get_instance()
 
     LogService.log().info(f'[batch valorant] (initialize): create some dirs if not exists')
     non_db = mongo_instance.get_db(f'{mongo.get("database").get("names")[0]}')
-    
-    r"""
+
+    """
     see:
         https://stackoverflow.com/questions/72323776/find-only-specific-key-from-mongodb-using-pymongo
         https://stackoverflow.com/questions/6570371/when-to-use-and-when-to-use-is
@@ -132,9 +117,9 @@ def initialize():
 
     LogService.log().info(f'[batch valorant] (initialize): check if list agents if bigger than zero')
     if len(agents_list) <= 0:
-        insert_data(VALORANT, 
+        insert_data(VALORANT,
                     build_model_valorant(
-                        getData(f'{valorant.get("url")}/{AGENTS}')))
+                            get_data(f'{valorant.get("url")}/{AGENTS}')))
     else:
         LogService.log().info(f'[batch valorant] (initialize): some data founded doing validation')
         for agent in agents_list:
@@ -142,14 +127,14 @@ def initialize():
             if ag.hash_validator != agent.get('hash_validator'):
                 # make request to valorant to get new informations
                 try:
-                    insert_data(VALORANT, 
+                    insert_data(VALORANT,
                                 build_model_valorant(
-                                    getData(f'{valorant.get("url")}/{AGENTS}')))
+                                        get_data(f'{valorant.get("url")}/{AGENTS}')))
                 except Exception as e:
                     LogService.log().error(f'[batch valorant] (initialize): {str(e)}')
             else:
                 LogService.log().info(f'[batch valorant] (initialize): {AGENTS} already registered.')
 
 
-if __name__ =='__main__':
+if __name__ == '__main__':
     initialize()
