--- carrega_variaveis.py
@@ -1,51 +1,0 @@
-import os
-import re
-import yaml
-
-
-def parse_config(path, data=None, tag='${}'):
-    """
-    Load a yaml configuration file and resolve any environment variables
-    The environment variables must have !ENV before them and be in this format
-    to be parsed: ${VAR_NAME}.
-    E.g.:
-    database:
-        host: !ENV ${HOST}
-        port: !ENV ${PORT}
-    app:
-        log_path: !ENV '/var/${LOG_PATH}'
-        something_else: !ENV '${AWESOME_ENV_VAR}/var/${A_SECOND_AWESOME_VAR}'
-    :param str path: the path to the yaml file
-    :param str data: the yaml data itself as a stream
-    :param str tag: the tag to look for
-    :return: the dict configuration
-    :rtype: dict[str, T]
-    """
-    # pattern for global vars: look for ${word}
-    pattern = re.compile('.*?\${(\w+)}.*?')
-    loader = yaml.SafeLoader(open('services/env/build_env.py'))
-
-    # the tag will be used to mark where to start searching for the pattern
-    # e.g. somekey: !ENV somestring${MYENVVAR}blah blah blah
-    loader.add_implicit_resolver(tag, pattern, None)
-
-    def constructor_env_variables(loader ='application.yml', node='services/env/build_env.py'):
-        """
-        Extracts the environment variable from the node's value
-        :param yaml.Loader loader: the yaml loader
-        :param node: the current node in the yaml
-        :return: the parsed string that contains the value of the environment
-        variable
-        """
-        value = loader.construct_scalar(node)
-        match = pattern.findall(value)  # to find all env variables in line
-        if match:
-            full_value = value
-            for g in match:
-                full_value = full_value.replace(
-                    f'${{{g}}}', os.environ.get(g, g)
-                )
-            return full_value
-        return value
-
-    loader.add_constructor(tag, constructor_env_variables)
