+++ domain/valorant/model/base_model.py
@@ -1,8 +1,11 @@
 # -*- coding: utf-8 -*-
 
 import datetime
-import pytz
 import hashlib
+import os
+
+import pytz
+
 
 class BaseModel:
     """
@@ -10,7 +13,7 @@ class BaseModel:
         Class to keep out some methods and variables. The domain valorant has some releative
         and each path need to be classifier
     """
+
     def __init__(self, data: str) -> None:
-        self.created_at = datetime.datetime.now(pytz.timezone('America/Sao_Paulo')).strftime('%Y-%m-%dT%H:%M:%S')
+        self.created_at = datetime.datetime.now(pytz.timezone(os.environ.get("TZ"))).strftime('%Y-%m-%dT%H:%M:%S')
         self.hash_validator = hashlib.sha256(data.encode('utf-8')).hexdigest()
-
