+++ services/log/log_service.py
@@ -5,7 +5,8 @@ import logging
 
 class LogService:
     logging.basicConfig(level=logging.DEBUG,
-                        format='%(asctime)s %(levelname)s %(name)s %(threadName)s : %(message)s')
+                        format='%(asctime)s %(levelname)s: %(message)s')
+
     @staticmethod
     def log():
         return logging
